Game started
It is possible to start and play game. Hopefully I have fixed player's bodies, so drawing and erasing players should be OK now. I could have done a simplier version, but with double memory consumation. The problem was, that the index of body part is important information, but not included directly direct in body part. So when the body has been resized, index iformation could be lost because of wrapping. I could store the index in body part, but it can be easily calculate.